Added profile viewer
The profile viewer will actually display stats of the user and allow
user to start a free type, accuracy type, set options for the profile,
and switch profile. It’s kind of the home screen of a user on a desktop
computer.

NOT COMPLETE. I have more work to do on this. For example, there are no
buttons.